requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
future==0.16.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.7.1
MarkupSafe==1.0
packaging==17.1
pip-review==1.0
pyparsing==2.2.0
python-dotenv==0.8.2
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
